requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
gast==0.2.0
grpcio==1.17.0
h5py==2.8.0
idna==2.7
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.4
numpy==1.15.4
pandas==0.23.4
plac==0.9.6
preshed==2.0.1
protobuf==3.6.1
pyreadline==2.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
regex==2018.1.10
requests==2.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
spacy==2.0.18
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
Theano==1.0.3
thinc==6.12.1
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
contractions==0.0.17
tweet-preprocessor==0.5.0
